codevoice-backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# File handling and multipart forms
python-multipart>=0.0.6
aiofiles>=23.2.0

# Environment and configuration
python-dotenv>=1.0.0
pydantic[email]>=2.4.0

# Audio processing and speech
faster-whisper>=0.10.0
soundfile>=0.12.1
webrtcvad>=2.0.10
pyttsx3>=2.90

# Scientific computing
numpy>=1.24.0
torchaudio>=2.1.0

# HTTP clients
httpx>=0.27.0
aiohttp>=3.8.0

# AI and ML APIs
groq>=0.8.0
google-generativeai>=0.8.0
openai>=1.0.0

# Redis async client for pub/sub
redis>=4.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.22.0
pytest-mock>=3.10.0
prometheus_client>=0.16.0

# Git integration
PyGithub>=2.1.0

# Authentication and security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Database (MongoDB)
motor>=3.3.0
pymongo>=4.5.0
beanie>=1.23.0
dnspython>=2.4.0
